changed logo image
I changed the logo image that was used in class and changed it to one of my themes logs which is studio ghibli, so a cat with a green and cream color scheme should be there now. it is also  a box and not a circle, not sure if that will need to be changed later on.